+++ src/main/java/wooteco/subway/contoller/SectionController.java
@@ -1,0 +1,30 @@
+package wooteco.subway.contoller;
+
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+import wooteco.subway.dto.SectionRequest;
+import wooteco.subway.service.SectionService;
+
+import javax.validation.Valid;
+
+@RestController
+@RequestMapping("/lines/{lineId}/sections")
+public class SectionController {
+    private final SectionService sectionService;
+
+    public SectionController(SectionService sectionService) {
+        this.sectionService = sectionService;
+    }
+
+    @PostMapping
+    public ResponseEntity<Void> createSection(@PathVariable Long lineId, @RequestBody @Valid SectionRequest sectionRequest) {
+        sectionService.insertSectionInLine(lineId, sectionRequest);
+        return ResponseEntity.noContent().build();
+    }
+
+    @DeleteMapping
+    public ResponseEntity<String> deleteStationInLine(@PathVariable Long lineId, @RequestParam Long stationId) {
+        sectionService.delete(lineId, stationId);
+        return ResponseEntity.ok().build();
+    }
+}
