+++ src/main/java/wooteco/subway/dto/LineUpdateRequest.java
@@ -1,0 +1,26 @@
+package wooteco.subway.dto;
+
+import javax.validation.constraints.NotBlank;
+
+public class LineUpdateRequest {
+    @NotBlank
+    private String name;
+    @NotBlank
+    private String color;
+
+    private LineUpdateRequest() {
+    }
+
+    private LineUpdateRequest(String name, String color) {
+        this.name = name;
+        this.color = color;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getColor() {
+        return color;
+    }
+}
