+++ src/main/java/wooteco/subway/dto/StationRequest.java
@@ -1,0 +1,19 @@
+package wooteco.subway.dto;
+
+import javax.validation.constraints.NotBlank;
+
+public class StationRequest {
+    @NotBlank
+    private String name;
+
+    private StationRequest() {
+    }
+
+    private StationRequest(String name) {
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
