+++ src/main/java/wooteco/subway/dto/StationResponse.java
@@ -1,0 +1,28 @@
+package wooteco.subway.dto;
+
+import wooteco.subway.domain.Station;
+
+public class StationResponse {
+    private Long id;
+    private String name;
+
+    private StationResponse() {
+    }
+
+    private StationResponse(Long id, String name) {
+        this.id = id;
+        this.name = name;
+    }
+
+    public static StationResponse of(Station station) {
+        return new StationResponse(station.getId(), station.getName());
+    }
+
+    public Long getId() {
+        return id;
+    }
+
+    public String getName() {
+        return name;
+    }
+}
