+++ src/main/java/wooteco/subway/exception/LineIllegalArgumentException.java
@@ -1,0 +1,7 @@
+package wooteco.subway.exception;
+
+public class LineIllegalArgumentException extends IllegalArgumentException {
+    public LineIllegalArgumentException(String errorMessage) {
+        super(errorMessage);
+    }
+}
