+++ src/main/java/wooteco/subway/exception/SectionIllegalArgumentException.java
@@ -1,0 +1,7 @@
+package wooteco.subway.exception;
+
+public class SectionIllegalArgumentException extends IllegalArgumentException {
+    public SectionIllegalArgumentException(String errorMessage) {
+        super(errorMessage);
+    }
+}
