+++ src/main/java/wooteco/subway/exception/StationIllegalArgumentException.java
@@ -1,0 +1,7 @@
+package wooteco.subway.exception;
+
+public class StationIllegalArgumentException extends IllegalArgumentException {
+    public StationIllegalArgumentException(String errorMessage) {
+        super(errorMessage);
+    }
+}
