--- src/main/java/wooteco/subway/line/controller/LineController.java
@@ -1,53 +1,0 @@
-package wooteco.subway.line.controller;
-
-import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-import wooteco.subway.line.dto.LineCreateRequest;
-import wooteco.subway.line.dto.LineResponse;
-import wooteco.subway.line.dto.LineUpdateRequest;
-import wooteco.subway.line.service.LineService;
-
-import javax.validation.Valid;
-import java.net.URI;
-import java.util.List;
-
-@RestController
-@RequestMapping("/lines")
-public class LineController {
-    private final LineService lineService;
-
-    public LineController(LineService lineService) {
-        this.lineService = lineService;
-    }
-
-    @PostMapping
-    public ResponseEntity<LineResponse> createLine(@RequestBody @Valid LineCreateRequest lineCreateRequest) {
-        LineResponse lineResponse = lineService.save(lineCreateRequest);
-        return ResponseEntity.created(URI.create("/lines/" + lineResponse.getId())).body(lineResponse);
-    }
-
-    @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<List<LineResponse>> showLines() {
-        List<LineResponse> lineResponses = lineService.findAll();
-        return ResponseEntity.ok(lineResponses);
-    }
-
-    @GetMapping("/{id}")
-    public ResponseEntity<LineResponse> showLine(@PathVariable Long id) {
-        LineResponse lineResponse = lineService.find(id);
-        return ResponseEntity.ok(lineResponse);
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<String> deleteLine(@PathVariable Long id) {
-        lineService.delete(id);
-        return ResponseEntity.noContent().build();
-    }
-
-    @PutMapping("/{id}")
-    public ResponseEntity<Void> updateLine(@PathVariable Long id, @RequestBody @Valid LineUpdateRequest lineUpdateRequest) {
-        lineService.update(id, lineUpdateRequest);
-        return ResponseEntity.ok().build();
-    }
-}
