--- src/main/java/wooteco/subway/line/dto/LineUpdateRequest.java
@@ -1,26 +1,0 @@
-package wooteco.subway.line.dto;
-
-import javax.validation.constraints.NotBlank;
-
-public class LineUpdateRequest {
-    @NotBlank
-    private String name;
-    @NotBlank
-    private String color;
-
-    private LineUpdateRequest() {
-    }
-
-    private LineUpdateRequest(String name, String color) {
-        this.name = name;
-        this.color = color;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public String getColor() {
-        return color;
-    }
-}
