--- src/main/java/wooteco/subway/line/exception/LineIllegalArgumentException.java
@@ -1,7 +1,0 @@
-package wooteco.subway.line.exception;
-
-public class LineIllegalArgumentException extends IllegalArgumentException {
-    public LineIllegalArgumentException(String errorMessage) {
-        super(errorMessage);
-    }
-}
