--- src/main/java/wooteco/subway/section/exception/SectionIllegalArgumentException.java
@@ -1,7 +1,0 @@
-package wooteco.subway.section.exception;
-
-public class SectionIllegalArgumentException extends IllegalArgumentException {
-    public SectionIllegalArgumentException(String errorMessage) {
-        super(errorMessage);
-    }
-}
