--- src/main/java/wooteco/subway/station/controller/StationController.java
@@ -1,40 +1,0 @@
-package wooteco.subway.station.controller;
-
-import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
-import wooteco.subway.station.dto.StationRequest;
-import wooteco.subway.station.dto.StationResponse;
-import wooteco.subway.station.service.StationService;
-
-import javax.validation.Valid;
-import java.net.URI;
-import java.util.List;
-
-@RestController
-@RequestMapping("/stations")
-public class StationController {
-    private final StationService stationService;
-
-    public StationController(StationService stationService) {
-        this.stationService = stationService;
-    }
-
-    @PostMapping
-    public ResponseEntity<StationResponse> createStation(@Valid @RequestBody StationRequest stationRequest) {
-        StationResponse stationResponse = stationService.save(stationRequest.getName());
-        return ResponseEntity.created(URI.create("/stations/" + stationResponse.getId())).body(stationResponse);
-    }
-
-    @GetMapping(produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<List<StationResponse>> showStations() {
-        List<StationResponse> stationResponses = stationService.findAll();
-        return ResponseEntity.ok(stationResponses);
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteStation(@PathVariable Long id) {
-        stationService.delete(id);
-        return ResponseEntity.noContent().build();
-    }
-}
