--- src/main/java/wooteco/subway/station/domain/Station.java
@@ -1,36 +1,0 @@
-package wooteco.subway.station.domain;
-
-public class Station {
-    private final Long id;
-    private final String name;
-
-    private Station(Long id, String name) {
-        this.id = id;
-        this.name = name;
-    }
-
-    public static Station from(String name) {
-        return new Station(null, name);
-    }
-
-    public static Station of(Long id, String name) {
-        return new Station(id, name);
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public boolean isSameId(Long id){
-        return this.id.equals(id);
-    }
-
-    public boolean isSameName(String name){
-        return this.name.equals(name);
-    }
-}
-
