--- src/main/java/wooteco/subway/station/dto/StationRequest.java
@@ -1,19 +1,0 @@
-package wooteco.subway.station.dto;
-
-import javax.validation.constraints.NotBlank;
-
-public class StationRequest {
-    @NotBlank
-    private String name;
-
-    private StationRequest() {
-    }
-
-    private StationRequest(String name) {
-        this.name = name;
-    }
-
-    public String getName() {
-        return name;
-    }
-}
