--- src/main/java/wooteco/subway/station/dto/StationResponse.java
@@ -1,28 +1,0 @@
-package wooteco.subway.station.dto;
-
-import wooteco.subway.station.domain.Station;
-
-public class StationResponse {
-    private Long id;
-    private String name;
-
-    private StationResponse() {
-    }
-
-    private StationResponse(Long id, String name) {
-        this.id = id;
-        this.name = name;
-    }
-
-    public static StationResponse of(Station station) {
-        return new StationResponse(station.getId(), station.getName());
-    }
-
-    public Long getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-}
