--- src/main/java/wooteco/subway/station/exception/StationIllegalArgumentException.java
@@ -1,7 +1,0 @@
-package wooteco.subway.station.exception;
-
-public class StationIllegalArgumentException extends IllegalArgumentException {
-    public StationIllegalArgumentException(String errorMessage) {
-        super(errorMessage);
-    }
-}
